% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.tac.clean-package.R
\docType{package}
\name{tacApp-package}
\alias{tacApp}
\alias{tacApp-package}
\title{tacApp: An Application to Assess Company Power}
\description{
An application to assess company power.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Authors:
\itemize{
  \item Souad Koliaï \email{souad@2degrees-investing.org}
  \item Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID})
  \item Alex Axthelm \email{alex@2degrees-investing.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
