% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unique_rowids}
\alias{unique_rowids}
\title{Unique \code{rowid} of the \code{valid} dataset that produces valid results}
\format{
An object of class \code{integer} of length 938.
}
\usage{
unique_rowids
}
\description{
Results are valid if they have at least one real or virtual category.
}
\keyword{datasets}
