% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{useful}
\alias{useful}
\title{Private data added to .gitignore}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11259 rows and 35 columns.
}
\usage{
useful
}
\description{
Private data added to .gitignore
}
\keyword{datasets}
